requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==2.0.5
django-filter==1.1.0
django-grappelli==2.11.1
django-import-export==1.0.1
django-suit==0.2.26
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.0
odfpy==1.3.6
openpyxl==2.5.3
Pillow==5.1.0
PyJWT==1.6.4
pytz==2018.4
PyYAML==3.12
requests==2.18.4
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
